% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{cbindlist}
\alias{cbindlist}
\title{Column-bind a list of objects}
\usage{
cbindlist(alist)
}
\arguments{
\item{alist}{List of objects with compatible row counts.}
}
\value{
The column-bound result.
}
\description{
Thin wrapper around `do.call(cbind, ...)`.
}
\examples{
cbindlist(list(a = 1:3, b = 4:6))

}
