% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeEnvironment.R
\name{closeEnvironment}
\alias{closeEnvironment}
\title{Gracefully Stop the Watchdog and Cluster}
\usage{
closeEnvironment(env_info, remove_watchdog_stopfile = TRUE)
}
\arguments{
\item{env_info}{A list as returned by `startEnvironment()`, containing `cluster`, `stop_file`, and `worker_pids`.}

\item{remove_watchdog_stopfile}{Boolean. If true, removes stopfile corresponding to this watchdog/environment. Default is TRUE.}
}
\value{
NULL (invisibly).
}
\description{
The `closeEnvironment()` function stops the watchdog by creating the stop file,
then uses `jstopCluster()` to kill the associated Rscript.exe processes.
}
\details{
This function is useful for a clean shutdown after your computations are finished.
It signals the watchdog script to stop by creating the specified stop file, and then
forcefully kills the worker processes to ensure a clean slate.

After calling `closeEnvironment()`, if you also wish to stop the parallel cluster, call:
\code{parallel::stopCluster(env_info$cluster)}.
}
\examples{
\dontrun{
env_info <- startEnvironment(num_cores = 2)

# ... run computations ...

closeEnvironment(env_info)
# Now the watchdog should stop and the processes are terminated.

parallel::stopCluster(env_info$cluster)
}

}
