% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{deleteNULL}
\alias{deleteNULL}
\title{Remove NULL elements from a list}
\usage{
deleteNULL(alist)
}
\arguments{
\item{alist}{A list.}
}
\value{
A list with all `NULL` entries removed.
}
\description{
Remove NULL elements from a list
}
\examples{
deleteNULL(list(a = 1, b = NULL, c = 3))

}
