% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_write_parquet.R
\name{.klimo_infer_spec_system}
\alias{.klimo_infer_spec_system}
\title{Infer spec system from presence of metric/imperial units columns}
\usage{
.klimo_infer_spec_system(spec_dt)
}
\description{
Infer spec system from presence of metric/imperial units columns
}
\keyword{internal}
