% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_synapser.R
\name{.syn_anno_pack}
\alias{.syn_anno_pack}
\title{Pack a named list into annotations2 schema}
\usage{
.syn_anno_pack(annotations)
}
\arguments{
\item{annotations}{Named list.}
}
\value{
List suitable for annotations2 endpoint.
}
\description{
Pack a named list into annotations2 schema
}
\keyword{internal}
