% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_synapser.R
\name{.syn_anno_type}
\alias{.syn_anno_type}
\title{Determine Synapse annotation type (annotations2)}
\usage{
.syn_anno_type(x)
}
\arguments{
\item{x}{Atomic vector.}
}
\value{
Type string: BOOLEAN/LONG/DOUBLE/STRING.
}
\description{
Determine Synapse annotation type (annotations2)
}
\keyword{internal}
