% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_synapser.R
\name{.syn_base_url}
\alias{.syn_base_url}
\title{Base URL for Synapse repository services}
\usage{
.syn_base_url()
}
\value{
A length-1 character URL.
}
\description{
Base URL for Synapse repository services
}
\keyword{internal}
