% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_synapser.R
\name{.syn_build_presigned_maps}
\alias{.syn_build_presigned_maps}
\title{Build maps from a presigned url/batch response}
\usage{
.syn_build_presigned_maps(presigned_resp)
}
\arguments{
\item{presigned_resp}{Parsed response from `.syn_multipart_presigned_urls()`.}
}
\value{
List with `url_by_part` and `hdr_by_part`.
}
\description{
Build maps from a presigned url/batch response
}
\keyword{internal}
