% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_synapser.R
\name{.syn_guess_content_type}
\alias{.syn_guess_content_type}
\title{Guess content type from file extension}
\usage{
.syn_guess_content_type(path)
}
\arguments{
\item{path}{Local file path.}
}
\value{
MIME type string.
}
\description{
Guess content type from file extension
}
\keyword{internal}
