% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_synapser.R
\name{.syn_is_scalar_chr}
\alias{.syn_is_scalar_chr}
\title{Test for scalar character}
\usage{
.syn_is_scalar_chr(x)
}
\arguments{
\item{x}{Object.}
}
\value{
Logical scalar.
}
\description{
Test for scalar character
}
\keyword{internal}
