% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_synapser.R
\name{.syn_multipart_presigned_urls}
\alias{.syn_multipart_presigned_urls}
\title{Get presigned URLs for multipart upload parts}
\usage{
.syn_multipart_presigned_urls(
  upload_id,
  part_numbers,
  token = NULL,
  verbose = FALSE,
  dry_run = FALSE
)
}
\arguments{
\item{upload_id}{Multipart upload id.}

\item{part_numbers}{Integer vector of part numbers.}

\item{token}{Optional Synapse token (PAT). If `NULL`, env vars are used.}

\item{verbose}{Logical; emit request/response progress messages.}

\item{dry_run}{Logical; do not execute the request, return a description instead.}
}
\value{
Parsed JSON list.
}
\description{
Get presigned URLs for multipart upload parts
}
\keyword{internal}
