% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_synapser.R
\name{.syn_norm_path}
\alias{.syn_norm_path}
\title{Normalize a remote path}
\usage{
.syn_norm_path(path)
}
\arguments{
\item{path}{Character scalar.}
}
\value{
Normalized character scalar.
}
\description{
Converts backslashes to slashes, collapses repeated slashes, trims leading/trailing
slashes.
}
\keyword{internal}
