% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_synapser.R
\name{.syn_pick_storage_location}
\alias{.syn_pick_storage_location}
\title{Pick a storageLocationId from uploadDestinationLocations}
\usage{
.syn_pick_storage_location(x)
}
\arguments{
\item{x}{Parsed response object.}
}
\value{
storageLocationId (scalar) or NULL.
}
\description{
Pick a storageLocationId from uploadDestinationLocations
}
\keyword{internal}
