% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_synapser.R
\name{.syn_put_presigned_part}
\alias{.syn_put_presigned_part}
\title{PUT a multipart part to a presigned URL}
\usage{
.syn_put_presigned_part(
  url,
  raw_bytes,
  signed_headers = list(),
  verbose = FALSE
)
}
\arguments{
\item{url}{Presigned URL.}

\item{raw_bytes}{Raw vector.}

\item{signed_headers}{Named list of headers required by Synapse (may be empty).}

\item{verbose}{Logical.}
}
\value{
httr2 response.
}
\description{
PUT a multipart part to a presigned URL
}
\keyword{internal}
