% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_synapser.R
\name{.syn_verbose}
\alias{.syn_verbose}
\title{Message helper}
\usage{
.syn_verbose(verbose, ...)
}
\arguments{
\item{verbose}{Logical.}

\item{...}{Passed to `message()`.}
}
\description{
Message helper
}
\keyword{internal}
