% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_read_parquet.R
\name{dp_meta_summary}
\alias{dp_meta_summary}
\title{Quick human summary of Klimo metadata for a parquet file}
\usage{
dp_meta_summary(path, max_show = 12L)
}
\description{
Quick human summary of Klimo metadata for a parquet file
}
