% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_parquet_v2.R
\name{dp_print_detect}
\alias{dp_print_detect}
\title{Print a Dataplane detection result}
\usage{
dp_print_detect(x)
}
\arguments{
\item{x}{Detection result list from [dp_detect()].}
}
\value{
The input `x` (invisibly).
}
\description{
Convenience printer for objects returned by [dp_detect()].
}
