% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_close.R
\name{env_close}
\alias{env_close}
\title{Stop a Cluster Created by `env_start()`}
\usage{
env_close(env_info, remove_watchdog_stopfile = TRUE)
}
\arguments{
\item{env_info}{A list returned by [env_start()].}

\item{remove_watchdog_stopfile}{If `TRUE`, remove an existing stop file.}
}
\value{
`NULL`, invisibly.
}
\description{
Gracefully stops the cluster and removes an existing temporary stop file.
}
\examples{
\dontrun{
env_info <- env_start(num_cores = 2)
env_close(env_info)
}
}
