% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{flatten_list_columns}
\alias{flatten_list_columns}
\title{Flatten list-columns into character columns}
\usage{
flatten_list_columns(dt)
}
\arguments{
\item{dt}{A data.frame / tibble / data.table-like object.}
}
\value{
An object of the same class as `dt`, with list-columns flattened.
}
\description{
For any list-column, converts each element to a single string:
\itemize{
  \item `NULL` -> `NA`
  \item length-1 -> that value
  \item length>1 -> comma-separated string
}
}
\examples{
x <- data.frame(a = 1:3, b = I(list("x", c("y","z"), NULL)))
flatten_list_columns(x)

}
