% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_parquet_geo.R
\name{gp_geoparquet_attach_geo_to_table}
\alias{gp_geoparquet_attach_geo_to_table}
\title{Attach GeoParquet metadata to an Arrow Table}
\usage{
gp_geoparquet_attach_geo_to_table(tbl, geo_meta, geo_key = "geo")
}
\arguments{
\item{tbl}{An Arrow Table.}

\item{geo_meta}{GeoParquet metadata list.}

\item{geo_key}{Metadata key to store geo JSON (default `"geo"`).}
}
\value{
An Arrow Table with updated schema metadata.
}
\description{
Adds a `geo` JSON metadata entry to an Arrow Table's schema metadata.
}
