% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_parquet_geo.R
\name{gp_geoparquet_geo_json}
\alias{gp_geoparquet_geo_json}
\title{Serialize GeoParquet metadata to JSON}
\usage{
gp_geoparquet_geo_json(geo_meta)
}
\arguments{
\item{geo_meta}{A metadata list created by [gp_geoparquet_build_geo_meta()].}
}
\value{
A JSON string.
}
\description{
Serialize GeoParquet metadata to JSON
}
