% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_parquet_geo.R
\name{gp_geoparquet_is_geoparquet}
\alias{gp_geoparquet_is_geoparquet}
\title{Test whether a Parquet file is GeoParquet}
\usage{
gp_geoparquet_is_geoparquet(path, geo_key = "geo")
}
\arguments{
\item{path}{Parquet file path.}

\item{geo_key}{Metadata key holding geo JSON (default `"geo"`).}
}
\value{
Logical.
}
\description{
Test whether a Parquet file is GeoParquet
}
