% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_read_parquet.R
\name{klimo_meta_summary}
\alias{klimo_meta_summary}
\title{Quick human summary of Klimo metadata for a parquet file}
\usage{
klimo_meta_summary(path, max_show = 12L)
}
\description{
Quick human summary of Klimo metadata for a parquet file
}
