% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{mergeDTlist}
\alias{mergeDTlist}
\title{Merge a list of data.frames/data.tables by keys}
\usage{
mergeDTlist(alist, by)
}
\arguments{
\item{alist}{List of data.frames / data.tables.}

\item{by}{Character vector of key columns passed to `merge()`.}
}
\value{
A merged data.frame/data.table (depends on inputs).
}
\description{
Iteratively merges a list of tabular objects using `merge(..., all = TRUE)`.
}
\examples{
a <- data.frame(id = 1:3, x = 11:13)
b <- data.frame(id = 2:4, y = 21:23)
mergeDTlist(list(a, b), by = "id")

}
