% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{s3_upload_file}
\alias{s3_upload_file}
\title{Save an R object to S3 (as RDS)}
\usage{
s3_upload_file(object, bucket, s3_path, temp_dir = tempdir())
}
\arguments{
\item{object}{R object to serialize.}

\item{bucket}{Character scalar. Bucket name.}

\item{s3_path}{Character scalar. Destination S3 key (path).}

\item{temp_dir}{Character scalar. Directory to place the temporary file.}
}
\value{
The result returned by `aws.s3::put_object()`.
}
\description{
Serializes an R object via `saveRDS()` into a temporary file, uploads it to S3,
then removes the temporary file.
}
\examples{
\dontrun{
res <- jput_s3(list(a = 1), bucket = "my-bucket", s3_path = "tmp/test.rds")
}

}
