% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_synapser.R
\name{syn_entity_relpath}
\alias{syn_entity_relpath}
\title{Get the relative path of an entity under a base container}
\usage{
syn_entity_relpath(
  base_id,
  entity_id,
  token = NULL,
  verbose = FALSE,
  dry_run = FALSE
)

syn_entity_relpath(
  base_id,
  entity_id,
  token = NULL,
  verbose = FALSE,
  dry_run = FALSE
)
}
\arguments{
\item{base_id}{Synapse ID of the base container (typically a Project).}

\item{entity_id}{Synapse ID of a descendant entity (Folder/File).}

\item{token}{Synapse PAT.}

\item{verbose}{Logical.}

\item{dry_run}{Logical.}
}
\value{
Character scalar path relative to base_id (no leading slash).

Character scalar path relative to base_id (no leading slash).
}
\description{
Example: base_project_id="syn64728425", entity_id="syn72246747"
might return "zz_smoketests/synapse_rw".

Example: base_project_id="syn64728425", entity_id="syn72246747"
might return "zz_smoketests/synapse_rw".
}
