% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_synapser.R
\name{syn_find_file_in_folder}
\alias{syn_find_file_in_folder}
\title{Find a file entity (by name) in a folder}
\usage{
syn_find_file_in_folder(
  folder_id,
  file_name,
  token = NULL,
  verbose = FALSE,
  dry_run = FALSE
)
}
\arguments{
\item{folder_id}{Synapse ID of Folder.}

\item{file_name}{Name of file entity to locate (exact match).}

\item{token}{Synapse PAT.}

\item{verbose}{Logical.}

\item{dry_run}{Logical.}
}
\value{
Synapse FileEntity ID (e.g., "syn123...").
}
\description{
Find a file entity (by name) in a folder
}
