% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_synapser.R
\name{syn_get_annotations}
\alias{syn_get_annotations}
\title{Get annotations2 for an entity}
\usage{
syn_get_annotations(entity_id, token = NULL, verbose = FALSE, dry_run = FALSE)
}
\arguments{
\item{entity_id}{synId.}

\item{token}{Optional token (PAT).}

\item{verbose}{Logical.}

\item{dry_run}{Logical.}
}
\value{
Parsed JSON list containing `etag` and `annotations`.
}
\description{
Get annotations2 for an entity
}
\examples{
\dontrun{
ann <- syn_get_annotations("synEntityId")
names(ann$annotations)
}
}
