% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_synapser.R
\name{syn_get_children}
\alias{syn_get_children}
\title{List children of a container (project or folder)}
\usage{
syn_get_children(
  parent_id,
  include_types = c("folder", "file", "project"),
  page_size = 100L,
  token = NULL,
  verbose = FALSE,
  dry_run = FALSE
)
}
\arguments{
\item{parent_id}{Synapse ID of a Project or Folder.}

\item{include_types}{Character vector of Synapse types to include.
Common values: "folder", "file", "table", "link", "entityview", "dockerrepo".}

\item{page_size}{Page size for pagination (Synapse uses tokens).}

\item{token}{Synapse PAT.}

\item{verbose}{Logical.}

\item{dry_run}{Logical.}
}
\value{
data.frame with columns: id, name, type
}
\description{
List children of a container (project or folder)
}
