% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_synapser.R
\name{syn_update_file_entity}
\alias{syn_update_file_entity}
\title{Update a FileEntity to point to a new file handle}
\usage{
syn_update_file_entity(
  entity_id,
  new_dataFileHandleId,
  token = NULL,
  verbose = FALSE,
  dry_run = FALSE
)
}
\arguments{
\item{entity_id}{FileEntity synId.}

\item{new_dataFileHandleId}{New file handle ID.}

\item{token}{Optional token (PAT).}

\item{verbose}{Logical.}

\item{dry_run}{Logical.}
}
\value{
Parsed JSON list for updated entity.
}
\description{
Update a FileEntity to point to a new file handle
}
\examples{
\dontrun{
syn_update_file_entity("synFileEntityId", new_dataFileHandleId = "999999")
}
}
