% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{to_camel_case}
\alias{to_camel_case}
\title{Convert strings to lowerCamelCase}
\usage{
to_camel_case(string)
}
\arguments{
\item{string}{Character vector.}
}
\value{
Character vector of the same length, converted to lowerCamelCase.
}
\description{
Converts arbitrary strings to lowerCamelCase by:
\enumerate{
  \item replacing non-alphanumeric characters with spaces
  \item splitting on whitespace
  \item lowercasing the first token
  \item TitleCasing subsequent tokens
  \item concatenating tokens
}
}
\examples{
to_camel_case(c("hello world", "some_value", "  already  camel  "))

}
