% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{trimallcols}
\alias{trimallcols}
\title{Trim whitespace from all character columns}
\usage{
trimallcols(data)
}
\arguments{
\item{data}{A data.frame or data.table.}
}
\value{
The trimmed object (copy).
}
\description{
Applies `trimws()` to all character columns in a data.frame/data.table.
}
\examples{
d <- data.frame(a = c("  x", "y  "), b = 1:2, stringsAsFactors = FALSE)
trimallcols(d)

}
