% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_close.R
\name{env_close}
\alias{env_close}
\title{Gracefully Stop the Watchdog and Cluster}
\usage{
env_close(env_info, remove_watchdog_stopfile = TRUE)
}
\arguments{
\item{env_info}{A list as returned by `env_start()`, containing `cluster`, `stop_file`, and `worker_pids`.}

\item{remove_watchdog_stopfile}{Boolean. If true, removes stopfile corresponding to this watchdog/environment. Default is TRUE.}
}
\value{
NULL (invisibly).
}
\description{
The `env_close()` function stops the watchdog by creating the stop file,
then uses `stop_cluster()` to kill the associated Rscript.exe processes.
}
\details{
This function is useful for a clean shutdown after your computations are finished.
It signals the watchdog script to stop by creating the specified stop file, and then
forcefully kills the worker processes to ensure a clean slate.

After calling `env_close()`, if you also wish to stop the parallel cluster, call:
\code{parallel::stopCluster(env_info$cluster)}.
}
\examples{
\dontrun{
env_info <- env_start(num_cores = 2)

# ... run computations ...

env_close(env_info)
# Now the watchdog should stop and the processes are terminated.

parallel::stopCluster(env_info$cluster)
}

}
