% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_parquet_v2.R
\name{print.dp_spec}
\alias{print.dp_spec}
\title{Print a Dataplane spec}
\usage{
\method{print}{dp_spec}(x, ...)
}
\arguments{
\item{x}{A `dp_spec`.}

\item{...}{Unused.}
}
\value{
The input `x` (invisibly).
}
\description{
Compact printer for `dp_spec` objects.
}
