% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_cluster.R
\name{stop_cluster}
\alias{stop_cluster}
\title{Forcefully Stop Worker Processes}
\usage{
stop_cluster(workerpids = NULL)
}
\arguments{
\item{workerpids}{Integer vector of worker PIDs. If `NULL`, kills all `Rscript.exe` (Windows)
or `R` (Linux) processes. Use with caution.}
}
\value{
NULL invisibly.
}
\description{
`stop_cluster()` attempts to kill the processes by PID.
On Windows, uses `taskkill /F /PID`.
On Linux, uses `kill -9`.
}
\examples{
\dontrun{
stop_cluster(env_info$worker_pids)
}

}
