% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimo_synapser.R
\name{syn_download_file_path}
\alias{syn_download_file_path}
\title{Download a file using a remote path (relative to a base container)}
\usage{
syn_download_file_path(
  base_id,
  remote_file_path,
  dest_path,
  overwrite = FALSE,
  token = NULL,
  verbose = FALSE,
  dry_run = FALSE
)
}
\arguments{
\item{base_id}{Base container synId.}

\item{remote_file_path}{Remote file path like `"a/b/c/file.csv"`.}

\item{dest_path}{Local destination path.}

\item{overwrite}{Logical; overwrite existing local file.}

\item{token}{Optional token (PAT).}

\item{verbose}{Logical.}

\item{dry_run}{Logical.}
}
\value{
`dest_path`.
}
\description{
Download a file using a remote path (relative to a base container)
}
\examples{
\dontrun{
syn_download_file_path(
  base_id = "synProjectId",
  remote_file_path = "zz_smoketests/synapse_rw/test2/New Text Document.txt",
  dest_path = "C:/tmp/downloads/test.txt",
  overwrite = TRUE,
  verbose = TRUE
)
}
}
